requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=/workspace/requirements.txt /workspace/requirements.in
#
build==1.4.0
    # via pip-tools
click==8.3.1
    # via pip-tools
coverage[toml]==7.13.4
    # via pytest-cov
exceptiongroup==1.3.1
    # via pytest
iniconfig==2.3.0
    # via pytest
packaging==26.0
    # via
    #   build
    #   pytest
    #   wheel
pip-tools==7.5.3
    # via -r /workspace/requirements.in
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pygments==2.19.2
    # via pytest
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==9.0.2
    # via
    #   -r /workspace/requirements.in
    #   pytest-cov
pytest-cov==7.0.0
    # via -r /workspace/requirements.in
pyyaml==6.0.3
    # via -r /workspace/requirements.in
ruff==0.15.2
    # via -r /workspace/requirements.in
tomli==2.4.0
    # via
    #   build
    #   coverage
    #   pip-tools
    #   pytest
typing-extensions==4.15.0
    # via exceptiongroup
wheel==0.46.3
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
